+++ inspect_wandb/models/hooks.py
@@ -206,7 +206,8 @@ class WandBModelHooks(Hooks):
         """
         if data.spec.metadata is None:
             return None
-        return { k[len("inspect_wandb_models_"):]: v for k,v in data.spec.metadata.items() if k.lower().startswith("inspect_wandb_models_")}
+        overrides = {k[len("inspect_wandb_models_"):]: v for k,v in data.spec.metadata.items() if k.lower().startswith("inspect_wandb_models_")}
+        return overrides
 
     def _load_settings(self, overrides: dict[str, Any] | None = None) -> None:
         if self.settings is None or overrides is not None: